requirements.txt
Flask>=0.12.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
psycopg2==2.7.1
Flask-Testing==0.6.2
